Fix: eslint warnings. Refactor email context imports and create custom hooks for better structure
